Provide some minimal replacements of mintlib headers

+++ tests/bench/sources/bench.c
@@ -171,7 +171,7 @@ int main(int argc, char *argv[])
 		(void)bench4(i & 0xf);
 	}
 	
-	printf("Executed in %ld seconds.\r\n", time(NULL) - start);
+	printf("Executed in %ld seconds.\r\n", (long)time(NULL) - start);
   } else {
 	printf("Provide number of iterations\r\n");
   }
